requirements.txt
setuptools>=58.0.0
Flask>=2.0.0
Werkzeug>=2.0.0
numpy>=1.20.0
scipy>=1.7.0
librosa>=0.8.0
matplotlib>=3.4.0
flask_wtf>=1.0.0
wtforms>=3.0.0
python-dotenv>=0.19.0
gunicorn>=20.0.0
pillow>=8.0.0
